+++ lib/data/dto/auth.dart
